+++ R/data.R
@@ -1,0 +1,14 @@
+#' parsnip model specification database
+#'
+#' This is used in the RStudio add-in and captures information about mode
+#' specifications in various R packages.
+#'
+#' @name model_db
+#' @aliases model_db
+#' @docType data
+#' @return \item{model_db}{a data frame}
+#' @keywords datasets internal
+#' @examples
+#' data(model_db)
+NULL
+