+++ R/descriptors.R
@@ -250,9 +250,9 @@ get_descr_spark <- function(formula, data) {
   .obs   <- function() obs
   .lvls  <- function() y_vals
   .facts <- function() factor_pred
-  .x       <- function() abort("Descriptor `.x()` not defined for Spark.")
-  .y       <- function() abort("Descriptor `.y()` not defined for Spark.")
-  .dat     <- function() abort("Descriptor `.dat()` not defined for Spark.")
+  .x       <- function() abort("Descriptor .x() not defined for Spark.")
+  .y       <- function() abort("Descriptor .y() not defined for Spark.")
+  .dat     <- function() abort("Descriptor .dat() not defined for Spark.")
 
   # still need .x(), .y(), .dat() ?