+++ R/logistic_reg_data.R
@@ -108,10 +108,11 @@ set_pred(
       res_1 <- res_2
       res_1$lo <- 1 - res_2$hi
       res_1$hi <- 1 - res_2$lo
-      res <- bind_cols(res_1, res_2)
       lo_nms <- paste0(".pred_lower_", object$lvl)
       hi_nms <- paste0(".pred_upper_", object$lvl)
-      colnames(res) <- c(lo_nms[1], hi_nms[1], lo_nms[2], hi_nms[2])
+      colnames(res_1) <- c(lo_nms[1], hi_nms[1])
+      colnames(res_2) <- c(lo_nms[2], hi_nms[2])
+      res <- bind_cols(res_1, res_2)
 
       if (object$spec$method$pred$conf_int$extras$std_error)
         res$.std_error <- results$se.fit
@@ -187,7 +188,7 @@ set_pred(
     args =
       list(
         object = expr(object$fit),
-        newx = expr(as.matrix(new_data[, rownames(object$fit$beta)])),
+        newx = expr(as.matrix(new_data[, rownames(object$fit$beta), drop = FALSE])),
         type = "response",
         s = expr(object$spec$args$penalty)
       )
@@ -509,22 +510,22 @@ set_pred(
       res_1 <- res_2
       res_1$lo <- 1 - res_2$hi
       res_1$hi <- 1 - res_2$lo
-      res <- bind_cols(res_1, res_2)
       lo_nms <- paste0(".pred_lower_", object$lvl)
       hi_nms <- paste0(".pred_upper_", object$lvl)
-      colnames(res) <- c(lo_nms[1], hi_nms[1], lo_nms[2], hi_nms[2])
+      colnames(res_1) <- c(lo_nms[1], hi_nms[1])
+      colnames(res_2) <- c(lo_nms[2], hi_nms[2])
+      res <- bind_cols(res_1, res_2)
 
-      if (object$spec$method$pred$conf_int$extras$std_error)
+      if (object$spec$method$pred$conf_int$extras$std_error) {
         res$.std_error <- apply(results, 2, sd, na.rm = TRUE)
+      }
       res
     },
-    func = c(pkg = "rstanarm", fun = "posterior_linpred"),
+    func = c(pkg = "parsnip", fun = "stan_conf_int"),
     args =
       list(
-        object = quote(object$fit),
-        newdata = quote(new_data),
-        transform = TRUE,
-        seed = expr(sample.int(10^5, 1))
+        object = expr(object$fit),
+        newdata = expr(new_data)
       )
   )
 )
@@ -554,10 +555,11 @@ set_pred(
       res_1 <- res_2
       res_1$lo <- 1 - res_2$hi
       res_1$hi <- 1 - res_2$lo
-      res <- bind_cols(res_1, res_2)
       lo_nms <- paste0(".pred_lower_", object$lvl)
       hi_nms <- paste0(".pred_upper_", object$lvl)
-      colnames(res) <- c(lo_nms[1], hi_nms[1], lo_nms[2], hi_nms[2])
+      colnames(res_1) <- c(lo_nms[1], hi_nms[1])
+      colnames(res_2) <- c(lo_nms[2], hi_nms[2])
+      res <- bind_cols(res_1, res_2)
 
       if (object$spec$method$pred$pred_int$extras$std_error)
         res$.std_error <- apply(results, 2, sd, na.rm = TRUE)