+++ tests/testthat/test_boost_tree_xgboost.R
@@ -159,8 +159,24 @@ test_that('xgboost regression prediction', {
 
   form_pred <- predict(form_fit$fit, newdata = xgb.DMatrix(data = as.matrix(mtcars[1:8, -1])))
   expect_equal(form_pred, predict(form_fit, new_data = mtcars[1:8, -1])$.pred)
+
+  expect_equal(form_fit$fit$params$objective, "reg:squarederror")
+
 })
 
+
+
+test_that('xgboost alternate objective', {
+  skip_if_not_installed("xgboost")
+
+  spec <-
+    boost_tree() %>%
+    set_engine("xgboost", objective = "reg:pseudohubererror") %>%
+    set_mode("regression")
+
+  xgb_fit <- spec %>% fit(mpg ~ ., data = mtcars)
+  expect_equal(xgb_fit$fit$params$objective, "reg:pseudohubererror")
+})
 
 
 test_that('submodel prediction', {
@@ -281,20 +297,20 @@ test_that('early stopping', {
     regex = NA
   )
 
- expect_warning(
+  expect_warning(
     reg_fit <-
       boost_tree(trees = 20, stop_iter = 30, mode = "regression") %>%
       set_engine("xgboost", validation = .1) %>%
       fit(mpg ~ ., data = mtcars[-(1:4), ]),
     regex = "`early_stop` was reduced to 19"
   )
- expect_error(
-   reg_fit <-
-     boost_tree(trees = 20, stop_iter = 0, mode = "regression") %>%
-     set_engine("xgboost", validation = .1) %>%
-     fit(mpg ~ ., data = mtcars[-(1:4), ]),
-   regex = "`early_stop` should be on"
- )
+  expect_error(
+    reg_fit <-
+      boost_tree(trees = 20, stop_iter = 0, mode = "regression") %>%
+      set_engine("xgboost", validation = .1) %>%
+      fit(mpg ~ ., data = mtcars[-(1:4), ]),
+    regex = "`early_stop` should be on"
+  )
 })
 
 
@@ -366,7 +382,7 @@ test_that('xgboost data and sparse matrices', {
 
 test_that('argument checks for data dimensions', {
 
-  skip_if_not_installed("C50")
+  skip_if_not_installed("xgboost")
 
   data(penguins, package = "modeldata")
   penguins <- na.omit(penguins)
@@ -379,9 +395,14 @@ test_that('argument checks for data dimensions', {
   penguins_dummy <- model.matrix(species ~ ., data = penguins)
   penguins_dummy <- as.data.frame(penguins_dummy[, -1])
 
-  f_fit  <- spec %>% fit(species ~ ., data = penguins)
-  xy_fit <- spec %>% fit_xy(x = penguins_dummy, y = penguins$species)
-
+  expect_warning(
+    f_fit  <- spec %>% fit(species ~ ., data = penguins),
+    "1000 samples were requested"
+  )
+  expect_warning(
+    xy_fit <- spec %>% fit_xy(x = penguins_dummy, y = penguins$species),
+    "1000 samples were requested"
+  )
   expect_equal(f_fit$fit$params$colsample_bytree, 1)
   expect_equal(f_fit$fit$params$min_child_weight, nrow(penguins))
   expect_equal(xy_fit$fit$params$colsample_bytree, 1)