+++ NEWS.md
@@ -1,3 +1,5 @@
+# parsnip (development version)
+
 # parsnip 0.1.5
 
 * An RStudio add-in is availble that makes writing multiple `parsnip` model specifications to the source window. It can be accessed via the IDE addin menus or by calling `parsnip_addin()`.